requirements.txt
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
google-genai==0.5.0
python-dotenv==1.0.1
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
